+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Instead of a LICENSE.txt file one can use a LICENSE-HEADER.txt file for a shortened license
+
 ## 0.16.4 - 2024-04-22
 ### Changed
 - Update nodejs to 21.6.2